update definition for BNGetAddressRenderedWidth()

+++ binaryninjacore.h
@@ -4867,7 +4867,7 @@ __attribute__ ((format (printf, 1, 2)))
 	BINARYNINJACOREAPI BNMemoryUsageInfo* BNGetMemoryUsageInfo(size_t* count);
 	BINARYNINJACOREAPI void BNFreeMemoryUsageInfo(BNMemoryUsageInfo* info, size_t count);
 
-	BINARYNINJACOREAPI uint32_t BNGetAddressRenderedWidth(size_t addr);
+	BINARYNINJACOREAPI uint32_t BNGetAddressRenderedWidth(uint64_t addr);
 
 #ifdef __cplusplus
 }
